+++ app/assets/javascripts/helpers/markdown_editor.js
@@ -130,6 +130,14 @@ Diaspora.MarkdownEditor.prototype = {
     }
   },
 
+  isPreviewMode: function() {
+    return this.instance !== undefined && this.instance.$editor.find(".md-preview").length > 0;
+  },
+
+  userInputEmpty: function() {
+    return this.instance === undefined || this.instance.getContent().length === 0;
+  },
+
   localize: function() {
     var locale = Diaspora.I18n.language;
 
@@ -159,4 +167,8 @@ Diaspora.MarkdownEditor.prototype = {
 
     return locale;
   }
+};
+
+Diaspora.MarkdownEditor.simplePreview = function($mdInstance) {
+  return "<div class='preview-content'>" + app.helpers.textFormatter($mdInstance.getContent()) + "</div>";
 };